# Conflicts:
#	public/js/gestionEvent.js

+++ src/Controller/GroupeYamlFileRepertoireController.php
@@ -4,9 +4,11 @@ namespace App\Controller;
 
 use App\Entity\Groupe;
 use App\Entity\GroupeYamlFileRepertoire;
+use App\Entity\Repertoire;
 use App\Entity\Utilisateur;
 use App\Entity\UtilisateurYamlFileRepertoire;
 use App\Entity\YamlFile;
+use App\Form\DirectoryGroupeType;
 use App\Form\GroupeYamlFileRepertoireType;
 use App\Form\PartagerYamlFileGroupeType;
 use App\Repository\YamlFileRepository;
@@ -16,6 +18,7 @@ use DomainException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -31,20 +34,90 @@ final class GroupeYamlFileRepertoireController extends AbstractController
         Groupe $groupe,
         Request $request,
         EntityManagerInterface $entityManager,
-        FlashMessageHelperInterface $flashMessageHelper
     ): Response {
-        $utilisateur = $this->getUser();
+        $repertoireRepository = $entityManager->getRepository(Repertoire::class);
+        $gyrRepository = $entityManager->getRepository(GroupeYamlFileRepertoire::class);
 
-        $formImport = $this->createForm(GroupeYamlFileRepertoireType::class);
+        $repertoire = new Repertoire();
+        $form = $this->createForm(DirectoryGroupeType::class, $repertoire, [
+            'groupe' => $groupe,
+        ]);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Associer l'utilisateur au répertoire
+            $repertoire->setGroupeRepertoire($groupe);
+
+            // Si aucun parent n'est sélectionné, utiliser le répertoire racine
+            if ($repertoire->getParent() === null) {
+                $repertoireRacine = $repertoireRepository->recupererRepertoireRacineGroupe($groupe->getId());
+
+                if ($repertoireRacine) {
+                    $repertoire->setParent($repertoireRacine);
+                }
+            }
+
+            $entityManager->persist($repertoire);
+            $entityManager->flush();
+
+            $this->addFlash('success', 'Répertoire créé avec succès !');
+            return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+        }
+
+        // Récupérer le répertoire racine
+        $repertoireRacine = $repertoireRepository->recupererRepertoireRacineGroupe($groupe->getId());
+
+        // Récupérer tous les fichiers de l'utilisateur
+        $listGyr = $gyrRepository->recuperertoutYamlfileGroupeParRepertoire($groupe->getId());
+
+        return $this->render('yaml_file/listeYamlFileGroupe.html.twig', [
+            'groupe' => $groupe,
+            'listGyr' => $listGyr,
+            'repertoireRacine' => $repertoireRacine,
+            'formRepertoire' => $form
+        ]);
+    }
+
+    #[IsGranted(attribute: 'GROUPE_VIEW', subject: 'groupe')]
+    #[Route('/groupe/{id}/upload_fichier', name: 'upload_yaml_groupe', methods:  ['GET', 'POST'])]
+    public function uploadYamlFileGroupe(
+        Groupe $groupe,
+        EntityManagerInterface $entityManager,
+        Request $request,
+        FlashMessageHelperInterface $flashMessageHelper,
+    ) : Response {
+
+        $utilisateur = $this->getUser();
+        $repertoireRepository = $entityManager->getRepository(Repertoire::class);
+        $gyrRepository = $entityManager->getRepository(GroupeYamlFileRepertoire::class);
+
+        $gyr =  new GroupeYamlFileRepertoire();
+
+        $formImport = $this->createForm(GroupeYamlFileRepertoireType::class, $gyr, [
+            'groupe' => $groupe,
+        ]);
+
         $formImport->handleRequest($request);
 
         if ($formImport->isSubmitted() && $formImport->isValid()) {
             $uploadedFile = $formImport->get('yamlFile')->getData();
             $droit = $formImport->get('droit')->getData();
+            $repertoireId = $formImport->get('repertoire')->getData();
 
             if (!$uploadedFile) {
                 $this->addFlash('error', 'Aucun fichier reçu.');
-                return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+                return $this->redirectToRoute('upload_yaml_groupe', ['id' => $groupe->getId()]);
+            }
+
+            $nameFile = $uploadedFile->getClientOriginalName();
+            $exists = $gyrRepository->existsYamlFileGroupe($groupe->getId(), $nameFile, $repertoireId);
+
+            if ($exists) {
+                $this->addFlash('error', sprintf(
+                    'Un fichier nommé "%s" existe déjà pour votre groupe dans ce répertoire.',
+                    $nameFile
+                ));
+                return $this->redirectToRoute('upload_yaml_groupe', ['id' => $groupe->getId()]);
             }
 
             try {
@@ -56,32 +129,33 @@ final class GroupeYamlFileRepertoireController extends AbstractController
                 $content = file_get_contents($uploadedFile->getRealPath());
                 $yamlFile->assertNotEmpty($content);
 
+                $repertoire = $repertoireRepository->find($repertoireId);
+
                 $yamlFile->setNameFile($uploadedFile->getClientOriginalName());
                 $yamlFile->setBodyFile($content);
                 $yamlFile->setUtilisateurYamlfile($utilisateur);
 
-                // Création du YamlFileGroupe directement
-                $gyr = new GroupeYamlFileRepertoire();
-
                 $gyr->setDroit($droit);
                 $gyr->setGroupe($groupe);
+                $gyr->setRepertoire($repertoire);
                 $gyr->setYamlFile($yamlFile);
-//                $gyr->setRepertoire(null);
 
                 $entityManager->persist($yamlFile);
                 $entityManager->persist($gyr);
                 $entityManager->flush();
 
+                $repertoireNom = $repertoire->getFullPath();
+
                 $this->addFlash('success', sprintf(
-                    'Fichier "%s" ajouté avec succès au groupe "%s".',
+                    'Fichier "%s" ajouté avec succès au groupe "%s" dans "%s".',
                     $uploadedFile->getClientOriginalName(),
-                    $groupe->getNom()
+                    $groupe->getNom(),
+                    $repertoireNom
                 ));
 
-                return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+                return $this->redirectToRoute('upload_yaml_groupe', ['id' => $groupe->getId()]);
             } catch (DomainException $e) {
                 $this->addFlash('error', $e->getMessage());
-                return $this->redirectToRoute('yaml_upload');
             } catch (FileException $e) {
                 $this->addFlash('error', 'Erreur lors de la lecture du fichier YAML.');
             }
@@ -89,18 +163,36 @@ final class GroupeYamlFileRepertoireController extends AbstractController
 
         $flashMessageHelper->addFormErrorsAsFlash($formImport);
 
+        return $this->render('yaml_file/upload_groupe.html.twig', [
+            'formImport' => $formImport,
+            'groupe' => $groupe,
+        ]);
+    }
+
+    #[IsGranted(attribute: 'GROUPE_VIEW', subject: 'groupe')]
+    #[Route('/groupe/{id}/partager_fichier', name: 'partager_yaml_groupe', methods:  ['GET', 'POST'])]
+    public function partagerYamlGroupe(
+        Groupe $groupe,
+        Request $request,
+        FlashMessageHelperInterface $flashMessageHelper,
+        EntityManagerInterface $entityManager
+    ) : Response
+    {
+        $utilisateur = $this->getUser();
+
         $yamlFilesUtilisateur = $entityManager
             ->getRepository(UtilisateurYamlFileRepertoire::class)
             ->findYamlFilesForUser($utilisateur);
 
         $yamlChoices = [];
 
+        $gyrRepository = $entityManager->getRepository(GroupeYamlFileRepertoire::class);
+
         foreach ($yamlFilesUtilisateur as $uyfr) {
 
             $file = $uyfr->getYamlFile();
 
-            $existant = $entityManager->getRepository(GroupeYamlFileRepertoire::class)
-                ->findOneBy(['yamlFile' => $file, 'groupe' => $groupe]);
+            $existant = $gyrRepository->findOneBy(['yamlFile' => $file, 'groupe' => $groupe]);
 
             if (!$existant) {
                 $repertoire = $uyfr->getRepertoire();
@@ -117,6 +209,7 @@ final class GroupeYamlFileRepertoireController extends AbstractController
 
         $formExistant = $this->createForm(PartagerYamlFileGroupeType::class, null, [
             'yaml_choices' => $yamlChoices,
+            'groupe' => $groupe,
         ]);
 
         $formExistant->handleRequest($request);
@@ -126,65 +219,81 @@ final class GroupeYamlFileRepertoireController extends AbstractController
             $data = $formExistant->getData();
             $yamlId = $data['yamlId'];
             $droit = $data['droit'];
+            $repertoireId = $formExistant->get('repertoire')->getData();
 
             $yamlFile = $entityManager->getRepository(YamlFile::class)->find($yamlId);
 
             if (!$yamlFile) {
                 $this->addFlash('error', "Fichier introuvable.");
-                return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+                return $this->redirectToRoute('partager_yaml_groupe', ['id' => $groupe->getId()]);
             }
 
-            $existant = $entityManager->getRepository(GroupeYamlFileRepertoire::class)
-                ->findOneBy(['yamlFile' => $yamlFile, 'groupe' => $groupe]);
+            $exists = $gyrRepository->existsYamlFileGroupe($groupe->getId(), $yamlFile->getNameFile(), $repertoireId);
 
-            if ($existant) {
-                $this->addFlash('error', 'Ce fichier est déjà dans le groupe.');
-                return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+            if ($exists) {
+                $this->addFlash('error', sprintf(
+                    'Un fichier nommé "%s" existe déjà pour votre groupe dans ce répertoire.',
+                    $yamlFile->getNameFile()
+                ));
+                return $this->redirectToRoute('partager_yaml_groupe', ['id' => $groupe->getId()]);
             }
 
+            $repertoireRepository = $entityManager->getRepository(Repertoire::class);
+            $repertoire = $repertoireRepository->find($repertoireId);
 
-            $gyr = new GroupeYamlFileRepertoire();
+            $gyr =  new GroupeYamlFileRepertoire();
             $gyr->setYamlFile($yamlFile);
             $gyr->setGroupe($groupe);
             $gyr->setDroit($droit);
+            $gyr->setRepertoire($repertoire);
 
             $entityManager->persist($gyr);
             $entityManager->flush();
 
             $this->addFlash('success', 'Fichier partagé au groupe.');
-            return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+            return $this->redirectToRoute('partager_yaml_groupe', ['id' => $groupe->getId()]);
         }
 
+        $flashMessageHelper->addFormErrorsAsFlash($formExistant);
 
-        $fichiersGroupe = $entityManager
-            ->getRepository(GroupeYamlFileRepertoire::class)
-            ->findBy(['groupe' => $groupe]);
-
-        return $this->render('yaml_file/listeYamlFileGroupe.html.twig', [
+        return $this->render('yaml_file/partagerYamlFileGroupe.html.twig', [
             'groupe' => $groupe,
-            'fichiersGroupe' => $fichiersGroupe,
-            'formImport' => $formImport,
             'formExistant' => $formExistant,
         ]);
     }
 
-    #[Route('/groupe/{id}/supprimer-fichier/{yamlId}', name: 'supprimer_yaml_groupe', methods: ['POST'])]
+    #[Route('/groupe/{id}/supprimer_fichier/{yamlId}', name: 'deleteYamlFileGroupe', options: ["expose" => true], methods: ['DELETE'])]
     public function supprimerYamlDuGroupe(
         Groupe $groupe,
         int $yamlId,
-        EntityManagerInterface $entityManager
+        EntityManagerInterface $entityManager,
+        Request $request,
     ): Response {
 
-        $gyr = $entityManager->getRepository(GroupeYamlFileRepertoire::class)->findByYamlFileAndGroupe($yamlId, $groupe->getId());
+        $yamlFile = $entityManager->getRepository(YamlFile::class)->findOneBy(['id' =>  $yamlId]);
+
+        if (!$yamlFile) {
+            return new JsonResponse(null, Response::HTTP_NOT_FOUND);
+        }
+
+        $gyrRepository = $entityManager->getRepository(GroupeYamlFileRepertoire::class);
+
+        $gyr = $gyrRepository->findByYamlFileAndGroupe($yamlFile->getId(), $groupe->getId());
 
         //Gestion droit avec un voter, pas avec IsGranted car GroupeYamlfileRepertoire possède plusieurs clés primaires, et symphony ne peut traiter ce cas avec IsGranted.
         $this->denyAccessUnlessGranted('GROUPE_FILE_DELETE', $gyr);
 
+        $submittedToken = $request->getPayload()->get('_token');
+
+        if (!$this->isCsrfTokenValid('delete' . $groupe->getId() . $yamlFile->getId(), $submittedToken)) {
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
+        }
+
+        $gyrRepository->supprimerYamlfileGroupeParRepertoire($yamlFile->getId());
         $entityManager->remove($gyr);
         $entityManager->flush();
 
-        $this->addFlash('success', 'Fichier supprimé du groupe.');
-        return $this->redirectToRoute('fichiers_groupe', ['id' => $groupe->getId()]);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 
     #[Route('/groupe/{id}/modifier-fichier/{yamlId}', name: 'modifier_yaml_file_groupe', methods: ['GET', 'POST'])]
